Use position controller for all slaves. Improve visuals

+++ Assets/Scripts/GameConnection.cs
@@ -55,9 +55,9 @@ public abstract class GameConnection
 
 	public abstract void Connect();
 
-	public void SetLocalValue( int objectID, int valueType, int valueIndex, float value ) 
+	public void SetLocalValue( byte objectID, byte valueType, int valueIndex, float value ) 
 	{
-		KeyValuePair<byte,byte> localKey = new KeyValuePair<byte,byte>( (byte) objectID, (byte) valueType );
+		KeyValuePair<byte,byte> localKey = new KeyValuePair<byte,byte>( objectID, valueType );
 
 		if( !localValues.ContainsKey( localKey ) ) localValues[ localKey ] = new float[ TYPE_VALUES_NUMBER ];
 
@@ -69,9 +69,9 @@ public abstract class GameConnection
 		}
 	}
 
-	public float GetRemoteValue( int objectID, int valueType, int valueIndex )
+	public float GetRemoteValue( byte objectID, byte valueType, int valueIndex )
 	{
-		KeyValuePair<byte,byte> remoteKey = new KeyValuePair<byte,byte>( (byte) objectID, (byte) valueType );
+		KeyValuePair<byte,byte> remoteKey = new KeyValuePair<byte,byte>( objectID, valueType );
 
 		if( remoteValues.ContainsKey( remoteKey ) ) return remoteValues[ remoteKey ][ valueIndex ];
 
@@ -135,9 +135,9 @@ public abstract class GameConnection
 
 	protected abstract bool ReceiveUpdateMessage();
 
-	public float GetNetworkDelay( int objectID ) 
+	public float GetNetworkDelay( byte objectID ) 
 	{ 
-		if( inputDelays.ContainsKey( (byte) objectID ) ) return inputDelays[ (byte) objectID ];
+		if( inputDelays.ContainsKey( objectID ) ) return inputDelays[ objectID ];
 
 		return 0.0f; 
 	}
